requirements.txt
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
frozendict==2.3.8
html5lib==1.1
idna==3.4
lxml==4.9.3
multitasking==0.0.11
numpy==1.26.1
pandas==2.1.2
peewee==3.17.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
webencodings==0.5.1
yfinance==0.2.31
